Create street.py
added script to solve a logic/statistics problem
(5 dice, you rolled 1,3,4,4,5. You have 2 rolls left. Is it worth it to roll only 1 dice, the one thats doubled, to roll for a single 2 to get a street, or would it be better to take the 4 and 1 to go for a 2 and 6 or 1 and 2?)